fix unsigned / signed comparison

+++ moveit_core/robot_model/test/test.cpp
@@ -90,14 +90,13 @@ TEST_F(LoadPlanningModelsPr2, Model)
   const std::vector<const moveit::core::JointModel*>& joints = robot_model->getJointModels();
   for (std::size_t i = 0; i < joints.size(); ++i)
   {
-    ASSERT_EQ(joints[i]->getJointIndex(), i);
+    ASSERT_EQ(joints[i]->getJointIndex(), static_cast<int>(i));
     ASSERT_EQ(robot_model->getJointModel(joints[i]->getName()), joints[i]);
   }
   const std::vector<const moveit::core::LinkModel*>& links = robot_model->getLinkModels();
   for (std::size_t i = 0; i < links.size(); ++i)
   {
-    ASSERT_EQ(links[i]->getLinkIndex(), i);
-    //    std::cout << joints[i]->getName() << std::endl;
+    ASSERT_EQ(links[i]->getLinkIndex(), static_cast<int>(i));
   }
   moveit::tools::Profiler::Status();
 }
